staticfiles/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
distlib==0.3.3
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.10.0
django-environ==0.8.1
django-heroku==0.3.1
django-phonenumber-field==6.0.0
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.3.2
gunicorn==20.1.0
phonenumberslite==8.12.39
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
tzdata==2021.5
whitenoise==6.0.0
